+++ server/api/src/server/jebt/user/index.ts
@@ -1,8 +1,8 @@
-import { supabaseJebt } from "../../supabase/jebt";
-import { Member } from "../../../types/jebt/board";
-import { PostgrestError } from "@supabase/supabase-js";
-import to from "await-to-js";
-import { checkMember } from "../board";
+import type { PostgrestError } from '@supabase/supabase-js';
+import to from 'await-to-js';
+import type { Member } from '../../../types/jebt/board';
+import { supabaseJebt } from '../../supabase/jebt';
+import { checkMember } from '../board';
 
 /**
  * ## 获取工作区成员列表和自己的权限信息
@@ -11,17 +11,14 @@ import { checkMember } from "../board";
  * @returns 工作区成员列表和自己的权限信息
  */
 export const getJebtUserList = async (workspaceId: string, userId: string) => {
-  const { data, error } = (await supabaseJebt
-    .from("member")
-    .select("*")
-    .eq("workspaceId", workspaceId)) as {
+  const { data, error } = (await supabaseJebt.from('member').select('*').eq('workspaceId', workspaceId)) as {
     data: Member[];
     error: PostgrestError | null;
   };
-  if (error) throw new Error("服务器错误");
-  if (data.length === 0) throw new Error("未找到工作区");
+  if (error) throw new Error('服务器错误');
+  if (data.length === 0) throw new Error('未找到工作区');
   const user = data.find((item) => item.userId === userId);
-  if (!user) throw new Error("无权限");
+  if (!user) throw new Error('无权限');
   return { data, user };
 };
 
@@ -40,20 +37,15 @@ export const updateJebtUserRole = async ({
 }: {
   workspaceId: string;
   userId: string;
-  role: "admin" | "member";
+  role: 'admin' | 'member';
   currentUserId: string;
 }) => {
   // 检查用户权限
   const [error, _] = await to(checkMember(currentUserId, workspaceId));
   if (error) throw new Error(error.message);
-  const { data, error: memberError } = await supabaseJebt
-    .from("member")
-    .update([{ role }])
-    .eq("userId", userId)
-    .eq("workspaceId", workspaceId)
-    .select("*");
-  if (memberError) throw new Error("服务器错误");
-  if (!data) throw new Error("未找到用户");
+  const { data, error: memberError } = await supabaseJebt.from('member').update([{ role }]).eq('userId', userId).eq('workspaceId', workspaceId).select('*');
+  if (memberError) throw new Error('服务器错误');
+  if (!data) throw new Error('未找到用户');
   return data[0];
 };
 
@@ -78,11 +70,7 @@ export const deleteJebtUser = async ({
     const [error, _] = await to(checkMember(currentUserId, workspaceId));
     if (error) throw new Error(error.message);
   }
-  const { error: memberError } = await supabaseJebt
-    .from("member")
-    .delete()
-    .eq("userId", userId)
-    .eq("workspaceId", workspaceId);
-  if (memberError) throw new Error("服务器错误");
+  const { error: memberError } = await supabaseJebt.from('member').delete().eq('userId', userId).eq('workspaceId', workspaceId);
+  if (memberError) throw new Error('服务器错误');
   return true;
 };
